+++ src/clj/backtype/storm/daemon/task.clj
@@ -1,6 +1,7 @@
 (ns backtype.storm.daemon.task
   (:use [backtype.storm.daemon common])
   (:use [backtype.storm bootstrap])
+  (:use [clojure.contrib.seq :only [positions]])
   (:import [java.util.concurrent ConcurrentLinkedQueue ConcurrentHashMap])
   (:import [backtype.storm.hooks ITaskHook])
   (:import [backtype.storm.hooks.info SpoutAckInfo SpoutFailInfo
@@ -9,48 +10,65 @@
 
 (bootstrap)
 
-(defn- mk-fields-grouper [^Fields out-fields ^Fields group-fields num-tasks]
-  (fn [^List values]
-    (mod (tuple/list-hash-code (.select out-fields group-fields values))
-         num-tasks)
-    ))
+(defn- mk-fields-grouper [^Fields out-fields ^Fields group-fields ^List target-tasks]
+  (let [num-tasks (count target-tasks)
+        task-getter (fn [i] (.get target-tasks i))]
+    (fn [^List values]
+      (-> (.select out-fields group-fields values)
+          tuple/list-hash-code
+          (mod num-tasks)
+          task-getter))))
 
-(defn- mk-custom-grouper [^CustomStreamGrouping grouping ^Fields out-fields num-tasks]
-  (.prepare grouping out-fields num-tasks)
+(defn- mk-shuffle-grouper [^List target-tasks]
+  (let [num-tasks (count target-tasks)
+        choices (rotating-random-range num-tasks)]
+    (fn [tuple]
+      (->> (acquire-random-range-id choices num-tasks)
+           (.get target-tasks)))))
+
+(defn- mk-custom-grouper [^CustomStreamGrouping grouping ^TopologyContext context ^Fields out-fields target-tasks]
+  (.prepare grouping context out-fields target-tasks)
   (fn [^List values]
-    (.taskIndices grouping values)
+    (.chooseTasks grouping values)
     ))
 
 (defn- mk-grouper
   "Returns a function that returns a vector of which task indices to send tuple to, or just a single task index."
-  [^Fields out-fields thrift-grouping num-tasks]
-  (let [random (Random.)]
+  [^TopologyContext context ^Fields out-fields thrift-grouping ^List target-tasks]
+  (let [num-tasks (count target-tasks)
+        random (Random.)
+        target-tasks (vec (sort target-tasks))]
     (condp = (thrift/grouping-type thrift-grouping)
       :fields
         (if (thrift/global-grouping? thrift-grouping)
           (fn [tuple]
             ;; It's possible for target to have multiple tasks if it reads multiple sources
-            0 )
+            (first target-tasks))
           (let [group-fields (Fields. (thrift/field-grouping thrift-grouping))]
-            (mk-fields-grouper out-fields group-fields num-tasks)
+            (mk-fields-grouper out-fields group-fields target-tasks)
             ))
       :all
-        (fn [tuple]
-          (range num-tasks))
+        (fn [tuple] target-tasks)
       :shuffle
-        (let [choices (rotating-random-range num-tasks)]
-          (fn [tuple]
-            (acquire-random-range-id choices num-tasks)
-            ))
+        (mk-shuffle-grouper target-tasks)
+      :local-or-shuffle
+        (let [same-tasks (set/intersection
+                           (set target-tasks)
+                           (set (.getThisWorkerTasks context)))]
+          (if-not (empty? same-tasks)
+            (mk-shuffle-grouper (vec same-tasks))
+            (mk-shuffle-grouper target-tasks)))
       :none
         (fn [tuple]
-          (mod (.nextInt random) num-tasks))
+          (let [i (mod (.nextInt random) num-tasks)]
+            (.get target-tasks i)
+            ))
       :custom-object
         (let [grouping (thrift/instantiate-java-object (.get_custom_object thrift-grouping))]
-          (mk-custom-grouper grouping out-fields num-tasks))
+          (mk-custom-grouper grouping context out-fields target-tasks))
       :custom-serialized
         (let [grouping (Utils/deserialize (.get_custom_serialized thrift-grouping))]
-          (mk-custom-grouper grouping out-fields num-tasks))
+          (mk-custom-grouper grouping context out-fields target-tasks))
       :direct
         :direct
       )))
@@ -77,9 +95,12 @@
     obj
     ))
 
+(defn- get-context-hooks [^TopologyContext context]
+  (.getHooks context))
+
 (defmacro apply-hooks [topology-context method-sym info-form]
   (let [hook-sym (with-meta (gensym "hook") {:tag 'backtype.storm.hooks.ITaskHook})]
-    `(let [hooks# (.getHooks ~topology-context)]
+    `(let [hooks# (get-context-hooks ~topology-context)]
        (when-not (empty? hooks#)
          (let [info# ~info-form]
            (doseq [~hook-sym hooks#]
@@ -88,7 +109,7 @@
 
 (defn outbound-components
   "Returns map of stream id to component id to grouper"
-  [^TopologyContext topology-context]
+  [^TopologyContext topology-context ^TopologyContext user-context]
   (let [output-groupings (clojurify-structure (.getThisTargets topology-context))]
      (into {}
        (for [[stream-id component->grouping] output-groupings
@@ -98,9 +119,10 @@
          [stream-id
           (into {}
                 (for [[component tgrouping] component->grouping]
-                  [component (mk-grouper out-fields
+                  [component (mk-grouper user-context
+                                         out-fields
                                          tgrouping
-                                         (count (.getComponentTasks topology-context component))
+                                         (.getComponentTasks topology-context component)
                                          )]
                   ))]))
     ))
@@ -155,6 +177,7 @@
         
         report-error-and-die (fn [error]
                                (report-error error)
+                               (apply-hooks user-context .error error)
                                (suicide-fn))
 
         ;; heartbeat ASAP so nimbus doesn't reassign
@@ -172,7 +195,7 @@
         _ (doseq [klass (storm-conf TOPOLOGY-AUTO-TASK-HOOKS)]
             (.addTaskHook user-context (-> klass Class/forName .newInstance)))
 
-        stream->component->grouper (outbound-components topology-context)
+        stream->component->grouper (outbound-components topology-context user-context)
         component->tasks (reverse-map task-info)
         ;; important it binds to virtual port before function returns
         puller (msg/bind mq-context storm-id task-id)
@@ -207,9 +230,7 @@
                                          (when (= :direct grouper)
                                            ;;  TODO: this is wrong, need to check how the stream was declared
                                            (throw (IllegalArgumentException. "Cannot do regular emit to direct stream")))
-                                         (let [tasks (component->tasks out-component)
-                                               indices (collectify (grouper values))]
-                                           (for [i indices] (tasks i))))
+                                         (collectify (grouper values)))
                                        (stream->component->grouper stream))]
                         (apply-hooks user-context .emit (EmitInfo. values stream out-tasks))
                         (when (emit-sampler)
@@ -274,6 +295,7 @@
                                 ^TopologyContext topology-context ^TopologyContext user-context
                                 task-stats report-error-fn]
   (let [wait-fn (fn [] @storm-active-atom)
+        last-active (atom false)
         task-id (.getThisTaskId topology-context)
         component-id (.getThisComponentId topology-context)
         max-spout-pending (storm-conf TOPOLOGY-MAX-SPOUT-PENDING)
@@ -341,10 +363,20 @@
            ))
        (if (or (not max-spout-pending)
                (< (.size pending) max-spout-pending))
-         (if (wait-fn)
-           (.nextTuple spout)
-           (Time/sleep 100))
-         ;; TODO: log that it's getting throttled
+         (if-let [active? (wait-fn)]
+           (do
+             (when-not @last-active
+               (reset! last-active true)
+               (log-message "Activating spout " component-id ":" task-id)
+               (.activate spout))
+             (.nextTuple spout))
+           (do
+             (when @last-active
+               (reset! last-active false)
+               (log-message "Deactivating spout " component-id ":" task-id)
+               (.deactivate spout))
+             ;; TODO: log that it's getting throttled
+             (Time/sleep 100)))
          ))
      (fn []
        (let [^bytes ser-msg (msg/recv puller)]