+++ src/jvm/backtype/storm/grouping/CustomStreamGrouping.java
@@ -1,8 +1,10 @@
 package backtype.storm.grouping;
 
+import backtype.storm.task.TopologyContext;
 import backtype.storm.tuple.Fields;
 import java.io.Serializable;
 import java.util.List;
+import java.util.Map;
 
 public interface CustomStreamGrouping extends Serializable {
     
@@ -12,16 +14,14 @@ public interface CustomStreamGrouping extends Serializable {
      * 
      * It also tells the grouping the metadata on the stream this grouping will be used on.
      */
-   void prepare(Fields outFields, int numTasks);
+   void prepare(TopologyContext context, Fields outFields, List<Integer> targetTasks);
     
    /**
      * This function implements a custom stream grouping. It takes in as input
      * the number of tasks in the target bolt in prepare and returns the
-     * indices of the tasks to send the tuple to. Each index must be in the range
-     * [0, numTargetTasks-1]
+     * tasks to send the tuples to.
      * 
-     * @param tuple the values to group on
-     * @param numTargetTasks the number of tasks in the target bolt
+     * @param values the values to group on
      */
-   List<Integer> taskIndices(List<Object> values); 
+   List<Integer> chooseTasks(List<Object> values); 
 }