+++ src/jvm/backtype/storm/hooks/BaseTaskHook.java
@@ -36,4 +36,8 @@ public class BaseTaskHook implements ITaskHook {
     @Override
     public void boltFail(BoltFailInfo info) {
     }
+
+    @Override
+    public void error(Throwable error) {
+    }
 }