+++ src/jvm/backtype/storm/transactional/partitioned/IOpaquePartitionedTransactionalSpout.java
@@ -10,7 +10,17 @@ import java.util.Map;
  * This defines a transactional spout which does *not* necessarily
  * replay the same batch every time it emits a batch for a transaction id.
  */
-public interface IOpaquePartitionedTransactionalSpout<T> extends IComponent {    
+public interface IOpaquePartitionedTransactionalSpout<T> extends IComponent {
+    public interface Coordinator {
+        /**
+         * Returns true if its ok to emit start a new transaction, false otherwise (will skip this transaction).
+         * 
+         * You should sleep here if you want a delay between asking for the next transaction (this will be called 
+         * repeatedly in a loop).
+         */
+        boolean isReady();
+    }
+    
     public interface Emitter<X> {
         /**
          * Emit a batch of tuples for a partition/transaction. 
@@ -24,4 +34,5 @@ public interface IOpaquePartitionedTransactionalSpout<T> extends IComponent {
     }
     
     Emitter<T> getEmitter(Map conf, TopologyContext context);     
+    Coordinator getCoordinator(Map conf, TopologyContext context);     
 }